+++ query-engine/driver-adapters/src/conversion/mod.rs
@@ -1,7 +1,9 @@
 pub(crate) mod js_arg;
+pub(crate) mod js_to_quaint;
 
 pub(crate) mod mysql;
 pub(crate) mod postgres;
 pub(crate) mod sqlite;
 
 pub use js_arg::JSArg;
+pub use js_to_quaint::*;
